requirement.txt
requests
dnspython3
netaddr
numpy
colorlog
termcolor
chardet
PySocks
faker